pre_appfs: kill containers early if using persistent storage
podman does not support moving its root directory, but we need that
for (simple) updates so we remove its state db.
Removing this file has the consequence of making any running container
unremovable, even after reboot, so cleanup is no longer possible.

The simplest solution is just to make sure no container runs, which means
containers are brought down for the duration of updates in this mode.
Print a warning to request using tmpfs mode for production which does
not have this problem as storage is then read-only

We can then on proceed ot remove bolt_state.db in the update script