scripts/rootfs: replace 'other_rootfs_uptodate' logic with timestamps
the 'other_rootfs_uptodate' logic was too complex: need to update
both filesystems if there is a change, complicated checks to clear
the flag in the script, and would not notice usage of persist_file.

Change the logic to use a timestamp in /etc/.rootfs_update_timestamp
instead and consider FS identicals if the timestamp is the same.
persist_file has also been updated to modify the timestamp, so the
fs will no longer be incorrectly considered uptodate after its use