mkimage version check: tighten check
swupdate seems to have no interest in handling versions with
fields > 65535 in 'old format', so tighten the checks down.
Also check for max signed int for other fields in semver versions
while we are here...

And print a warning if alphanum are mixed that the number part
will only be sorted alphabetically e.g. 'at2' > 'at10'