scripts: add locking to serialize updates
updates are serialized within a single swupdate process but we rely
on multiple services for: usb updates, run from url list, hawkbit,
or whatever else our users might want to run (e.g. command line)

As locking needs to survive between multiple processes and intermediate
steps (we take lock in pre script, and only normally clear it on post),
we need to rely on the existance of an external process - in our case
the parent swupdate process - so failures to release the lock because
of a failed step does not result in a total deadlock.

We can still get stuck updates if a failed step in one swupdate process
leads to no cleanup there and another swupdate process tries to install
a new update, but this hopefully won't happen for normal production
systems.
If this is a problem, the script could query the said swupdate process
to see if an update is running through the swupdate API.