examples: add factory_reset example
factory reset example just optimistically copies sd image into running filesystem
and hopes for the best.

Note it does not update uboot, but it does reset boot to partition 0
before writing disk to save an extra reset.
if you need to update uboot as well, swdesc_boot will not work because
the symlink is not setup (scripts don't run): removing the debug skip
script and running swdesc_boot normally could work, but it might update
the wrong uboot, so the only real solution would be to open-code the
flashing or making the symlink manually before running swdesc_boot