scripts/podman_cleanup: identify used images by ids
Trying to guess if image is in used based on tag simplification probably works,
but it's best to rely on image id and podman internal parsing (image inspect <tag>)
to avoid weird corner cases.

Note that this way we no longer untag unused tags, so dummy updates that reference
the same image id but add an extra tag will not lead to previous tag removal.
That probably won't be a problem, but if it is we can keep track of which tags are
used as well by guessing again like the old code did...