versions management: forbid using a component with different versions
the previous check was too lax and does not make sense now versions can
be made 'different' easily.
Note this also applies to versions within subboards, e.g. we do not support
installing two different versions of uboot on different boards in same swu,
the version must be the same