scripts/versions: fix board check when old version was not present
installing official swu with multiple 'boot' version on a system with no
bootloader (.. which shouldn't happen either...) would lead to incorrect
file parsing and duplicate the 'boot' line.

duplicated lines are kept forever after that, so it's a bit ugly, but
the parsing works even in case of duplicate so shouldn't be a problem in
practice.