mkswu: add NO_PRESERVE_FILES option
if a user manage their local modifications themselves in their own base_os the
swupdate_preserve_files mechanism is a nuisance and should have an option to
be disabled.
Setting NO_PRESERVE_FILES in desc file or mkswu.conf will do that.

Note that we eventually will want to handle the 'post' phase of
swupdate_preserve_files in a separate step immediately after base_os,
so making it a standalone option is better than having users rely on
DEBUG_SWDESC