versions: remove duplicate lines
duplicate lines shouldn't cause any problem, but might as well remove them
to be sure.
This only keeps the first version for a given component, if the old file had

multiple versions they will be ignored. The previous behaviour would use the
first version as well for scripts, and it's not clear which version swupdate
would retain, so at least we're now safe from such mismatch