version check: fix bad install-if different version check regex
The check would fail with the following error:
grep: Invalid range end
because hyphen (-) was used not at the end of a range, whereas man grep states:
"Finally, to include a literal - place it last."

The regex was also just not complete:
 - normal versions and semvers with extra numerical values had these plain ignored
(e.g. 1.2.3.4.5 = 1.2.3.4 = 1.2.3.4.8, 1.2.3.4-abc = 1.2.3-abc)
 - the + regex required a dash component, but just 1.2+a and 1.2+b were considered
equal as well.

Finally, add tests for these two patterns

Reported-by: Daisuke Mizobuchi <mizo@atmark-techno.com>