post_boot: encrypted boot: re-set mmc bootpart during post processing
writing an encrypted boot image trying to boot it will just hang forever
with possibly no means of switching.
This safety shouldn't be needed (it's only required if someone manually
ran abos-ctrl rollback), but at least would avoid rollback -> update -> hang
patterns.

How much we'd want to protect is up for debate, but it's hard to get access
to $ab in the swdesc_imxboot_enc execution block and this is better than
nothing