mkswu pre script: run as a self-extracting shell script
replace swdesc_exec_nochroot by a simple swdesc_script_nochroot for
prescript, where the script itself is a base script + archive
concatenated

This is in preparation to adding some environmental values to the script
that will extend the archive, which cannot be done easily with the
current way this is done: while gnu tar allows chaining tar and will
correctly "pop" them one at at time, busybox tar is not that subtle,
and creating a tar then modifying it with --append which would also work
is not practical given we do not want to recreate archives that don't
change.
(well, in this case it probably wouldn't really have mattered...)

While we are here, rename entry scripts to scripts_pre and scripts_post
for clarity: users have been confusing swupdate_post.sh by a script to
run to run swupdate, hopefully it should be clearer it's related to the
scripts dir.