podman logs cleanup: rename 'podman' logging alias to podman_info
Aliasing 'podman' to a logging function sounded practical at the
time but leads to weird errors as seen in commit a248774146ac
("scripts/appfs: fix podman wait commands")

Use an alias name that is more clear about what it does so plain
podman invocations work as intended.
Now the worst that can happen is some logs go to error level
instead of info, but the update logic will not be broken.

Change-Id: I3f84f3932a377c6999bda3181613adc0a97d817e