initial_setup: clarify certificate installation
initial_setup would not update user certificates the same way depending
on whether UPDATE_CERTS is set or not.

This can be confusing to debug when the user would not understand why
the swdesc_exec command we actually do run has no effect, if they just
change the file there.

This doesn't warrant a version update as the change is actually NOOP in
practice.

Change-Id: Ie39e4688216cc81525c95e33d779dde2c4392964