swdesc_boot: remove 'strings' dependency
strings was used to avoid binary input, but grep with -ao works with
both gnu grep and busybox grep so just grep directly.
We are guaranted not to get binary output from the regex.

Note using strings is faster on pathological input (e.g. 4MB of \0 on
busybox grep takes almost 10s on g4), but a real uboot only takes 10ms
and random data (encrypted image that would have forgotten to set their
version) 170ms, so it is simpler to not keep the alternative and just
use grep directly.

Change-Id: I9d0b050c62e9943093c66c33dd080f5ba9d03de9