hawkbit-compose: mysql: fix path for extra config
mysql extra configuration files are in /etc/mysql/conf.d and not
/etc/mysql/mysql.conf.d

Failure to set the path properly means that database created with older
configurations will have their locales set to latin1, and if japanese or
other non-latin1 character must be stored then a conversion should be
made e.g. connect directly to the mysql container, run mysql there with
the root password in .env, and run charset conversion commands as
described in https://stackoverflow.com/a/6115705

There is no other impact, so if latin1 is good enough then no
manipulation is needed.

Change-Id: Ic87964be09ff01e7e459988b208b93fcdc977766