scripts: force fs types for mount commands
trying to mount a non-valid partition prints scary exFAT/F2FS error
messages as linux tries to probe all available filesystem types.
Since we know what filesystems are valid and these aren't, just always
set -t for all mount commands.

ext4 and btrfs in particular do not print any message in dmesg if the
filesystem is not valid (which can be argued with in itself), so this
de-facto silences all warnings about it

Change-Id: Icaae43450cd25e970b84696421f1aed77390953e