scripts: do not mount B-side if upgrade_available is unset
We've removed the scary messages with a previous commit but it is more
correct to not try to mount the other partition if update isn't set.

This doesn't fix anything obvious, but e.g. would allow install to pass
if power dropped after sw-versions have been written to B-side and
before u-boot got switched, so is probably worth doing.

Change-Id: Ib6b39062c2943caba0346540934b235f58d27880