scripts/update_shadow: only copy from persisted shadow
The normal process for baseos initialization is to first set a
password in overlay on login, which is not preserved unless the
user calls 'persist_file' on /etc/shadow.

When a user installs their initial_setup swu, there must be a password
set inside the swu and that password is successfully preserved, but for
pre-sale production it is also possible to install a SWU with public
key and 'ALLOW_EMPTY_LOGIN' that should not modify the install from the
default expired empty password.

In this case, if an operator first logs in and sets password in memory
we used to copy it an incorrectly set a password on a future user's
armadillo, which is bad.

To make this less likely, only look at password in the actual rootfs
without going through the overlay.

Note we need a fallback for the very first armadillo release which did
not have /live/rootfs, as other places with the same comment.

Change-Id: Icf8204dc9c35115dcf9c0b4fbca895ecb1893ea5
Reported-by: Naomi TODORI <naomi.todori@atmark-techno.com>