scripts/lock: check for rebooting in try_lock
If we start a second update before the first one stops, it will
wait for the update to end by looping on try_lock.
If the first update triggered a reboot we want to make sure it
did not create a 'reboot lock'.

It's also important to check after we've gotten the lock, as the
finished update will create reboot marker before killing itself,
but we might get raced if we just check before locking.

Change-Id: I71f378c65a48d8861c1ecd95b6a7e0f88f1b4669