scripts: simplify version strings on lookup
swupdate does not actually handle trailing '.0's and leading '0's as
originally designed, the versions are identical (even for install_if
different) so we should treat them as such.

Preprocessing might be done in mkswu itself but that would require
updates of existing identical components (e.g. u-boot version on files
would change from 2020.04 to 2020.4 and be considered different by our
scripts...), so we need to handle it in scripts/versions.sh.

This commit:
- add tests for swupdate actual behaviour (how this was found...)
- make get_version simplify leading 0s and trailing .0s
- add tests for new get_version
- remove tests for version_higher that were incorrect, hopefully
should never exercise these now...

Change-Id: I7beee41a18c4afff3b01d7425a98b63bb756e435