mkswu: add --update-version
We're starting to have many scripts that want to bump a desc file's
version, might as well share that code and test it.

This version improves over existing ones in that it allows specifying
a base to use for version: this doubles as a check so users will not
inadvertently try to 'update' the version to something lower that will
not be installed.
It does not understand install_if different, though, so a force switch
might be useful later if required...

Change-Id: If906f38ccac63152a14d800372885b92ec6b825a