scripts/pre_rootfs: make btrfs the default if not found
This should have absolutely no impact as findmnt should always work.
The actual default change happened in build-rootfs official images,
as we do not want to force btrfs for people with a very old uboot.

We might want to review this in the future as this will leave people
with e.g. secureboot on ext4, but it is better than making the new
system not bootable if they use an old uboot..

Change-Id: Iaff0fba8bbbc58a08e70fe1f401db29727664503