version simplifications: normalize version at mkswu time
The original version simplification code was simplifying version on
device, but was incorrect in the if-different case when multiple dashes
are involved (swupdate falls back to string comparison on these while
mkswu would incorrectly also simplify leading zeroes there)

The algorithm is complex enough to warrant its own normalization
function, and that makes more sense to do once when building swu than
everytime the swu is installed.

There were two reasons the simplification was done on the device:
 - the initial idea was to preserve the input version (not simplified)
all the way to sw-versions, but given the surprising behaviour that can
happen if simplification is involved it's better to store the simplified
version, and that's already what the current code does.
 - doing simplification on mkswu side means that if non-simplified
versions have been stored in sw-versions and a simplified version is
sent in the swu, the comparison would consider the versions different
even when they are not.
For atmark-provided binaries, only boot images are involved and a new
release has already been made with the new code which will store a
simplified version, there is no problem on this side.

If we want to play it safe with user-provided versions, the on-device
simplification could be kept around a bit longer but unlike 'boot' there
is no fallback handling, so the worst that could happen is a useless
reboot where swupdate does not install anything: simplifying the code
seems like a better approach.

Tests have been adapted to test the mkswu normalize_version function,
and a couple of new tests for multiple dashes on different version have
been added.
(eval was changed to really eval at this point to allow compound
instructions)

Change-Id: Ia28dea2826bcaabbb1ad0be0bd398ab26abc99ce