examples/kernel_update_plain.install.sh: check that desc file is compatible
When calling kernel_update_plain.install.sh on an existing desc we try to
keep the old desc file and just bump the version, but we need the desc
file to be compatible for this.
For now just check for a comment appearing in the first compatible version
of the desc, but later updates should add a subversion like we do with
config or swupdate_preserve_files. We do not try to handle it at this point
as there is nothing to do except remove the file and adding it back.

Change-Id: If1e0930864ab5869ac9aae2f4b19cc6ff81fedba