swdesc_* commands: avoid considering variables from outside.
Introduce two eval stanza to control which variables are used when:
- MKSWU_declare_swdesc_cmd_vars_stanza initializes variables from
MKSWU_* into short local names
- MKSWU_export_swdesc_cmd_vars_stanza sets MKSWU_* from short local
names

This is a bit more cumbersome to use as a swdesc_* command that uses
another swdesc_* command now needs to explicitely call the export
stanza, but this allows finer control and should avoid name clashes
if a user decides to use some local variable in their desc files

For compatibility, component/version/install_if are handled as variables
because they were accessible outside of swdesc_option in the first
versions of mkswu; this prints a warning and this compatibility layer
will be removed on 2024/04 update.

Other variables were left without fallback: they were never meant to be
accessed directly, and this was never intended for use. A word will be
added in the changelog to make users aware of this.
(If they really want to, they will be able to use MKSWU_xx isntead but
this will not be documented)

Change-Id: If7a511f72676d8b69211025dc94cab7f7bcf5ce4