swdesc_boot: add 1kb pad if required for imx6/7 boot image
Our u-boot build scripts aren't consistent in how boot images
are produced. In particular, imx6ull and imx7d require the
uboot image to be at 1k offset in the mmcblkXbootY partition,
so we need to add that padding if the build script didn't
do it

Since we do not know what target the image is being built for,
look at the u-boot image header and pad accordingly.

This also allows giving and error if something we do not recognize
as a boot image is passed as argument, as that might brick the device

Change-Id: I9cb203dfa0769c556dedc15d784e2b8259ead48d