swdesc_*: add --description option to mask info message
Allow overriding default displayed command individually.

This commit is a bit big because we keep the original command around
even when that was specified for mkswu --show:
- escape command immediately, conf-escaped commands cannot contain
  newlines
- if newline is present keep original command in a comment instead
- fix show to parse for this comment as well. We don't re-use the
old '# swdesc_xx' format because 1/ it doesn't handle new lines and
2/ it wasn't a good idea to have something without a well identified
prefix (swdesc_.. might not always be enough)

Also add tests to back this all up.

Change-Id: If7564e9d2c7c8b597d37f9d3af246452ce5c52b2