scripts/post_rootfs_baseos: remove wifi-recover
Previous commit adding /target to files incorrectly added it for
wifi-recover: the intent was to enable the service if it was not present
before.
But wifi-recover never actually got distributed so we can just remove
this section.

Also silence a few more harmless error messages when files do not exist.

Change-Id: I68ae9c9fe1abed834ab00da0efa897cb524ca05b
Fixes: 895535753b26 ("scripts/post_rootfs_baseos: add missing /target prefixes")