mkswu --show: bad unescaping and output order
- the unescape code was too crude and did not properly preserve e.g.
raw '\n' strings, escaped as '\n' : an even number of backslashes
should not expand the new line
- the fix for this requires the non-posix gensub and it was a good time
to look at 'asorti' for file sorting, but it's just as simple to get the
file list from the SWU cpio itself: this is a bit slower but will
correctly re-order base_os first (internal MKSWU_install_first
variable), and will work with previously generated swu without
making it another incompatible change.
The other advantage is that busybox awk supports gensub but not asorti,
so we can keep testing this on alpine.

Change-Id: I5cd06124f742f52946d856bfa95687aeddec7686