scripts/podman_cleanup: do not fail update for pre-existing bogus config
If a config file points to an image that is not available (and is
autostart + pull never) then we currently fail update, as the update
will result in something that won't start properly.

But if the config was pre-existing e.g. something an user added through
persist_files then we should just let this go: just issue a warning
instead.

Change-Id: Ida17eef1965ba02e3d7da24308e2afcb2ff4dfb8