deprecation notice: disallow commands without version altogether
Up until now SWUs with two commands, one setting a version and another
having no version was allowed, but this behaviour is confusing as the
user might not expect the other command to always run.
Make that explicit by adding a warning along with a deprecation notice
that we will hard fail in the future.

Change-Id: Ie5c31aa24a7019847141c88cfbdbd64cde6e2f2b