swdesc_*: try to preserve file extension
Some commands like apk will fail if their target hasn't preserved their
extension.
There was two places where we mangle file names:
- in swdesc_exec itself, on non-trivial commands we regenerate a
filename with a unique name based on command content.
In this case just carry the last component around if it is short.
- in write_entry_stdout when compressing or encrypting the file.
Prepend zst or enc instead of appending it.

For swdesc_exec the only reason we do the rename is to allow multiple
identical commands, and we could skip the first rename e.g. only do
rename after we detect a collision, but due to compression/encryption
rename it is not obvious at this stage and there doesn't seem to be
much use of trying too hard.

A better solution would be to have swdesc_exec take a new parameter and
expand the file with the original name on device, but since that will
require a new swupdate version it will take time to be effective.

Reported-by: Akihito Irie <akihito.irie@atmark-techno.com>
Change-Id: I178722284906be39c94ed235ae839d091c74eaf3
Fixes: 3d384c082493 ("mkswu: allow multiple swdesc_exec on a file if different command")