scripts/pre_boot: uboot copy: do not trust versions
We were copying boot image from current to other target if versions
changed, but in case a partial or broken u-boot was installed in an
update that failed, or rollback happened this is not enough.

Be more thorough: just copy boot image all the time unless an update is
provided or content is identical.

For linux image in boot partition, only do this if boot_linux version
has been set, it is not required most of the time (and would fail on
systems with a smaller boot partition)

Link: https://armadillo.atmark-techno.com/forum/armadillo/15730
Change-Id: I0e24f890d020ef83abc85f01500784b57636fb42