swupdate_preserve_files: add CHOWN user file syntax
allow swupdate_preserve_files to specify directories to chown.
This can be useful for users which do not have a fixed id, yet
configuration files are copied e.g. abos-web-user's /etc/atmark/abos_web
directory

Note that this still needs to work on older verisons of baseos,
so we need to either ignore chown errors or check if the user
exists on target first -- we did the later which adds a bit
of code.
Wildcard splitting is also slightly hard to follow (needs to adjust
IFS all the time) and might not be needed, but it would be odd that
only chown does not do wildcards...

Change-Id: I2e96fb37429a5f8b5e25cc7306aa46a1191bed50