mkswu: check for coreutils in main script
'mkswu' itself depends on coreutils.

There is no reason to change that at this point, if someone wants to run
mkswu on armadillo they can just install coreutils.

This can be revisited later if appropriate, but as we do not package
mkswu for armadillo in the first place this is not a problem

Change-Id: I0ece1d03d7c3197f9885727fd68ca33364b20533