swdesc_tar/files --dest: allow /tmp and /var/tmp without --extra-os
/tmp and /var/tmp are mounted in /target and valid copy destination
(for e.g. tempotary files before command)
They shouldn't require extra-os as they do not modify the rootfs.

While here, add an extra / at the end of the pattern in the case so
we can match '/foo/*' instead of /foo* as previously, while that
was somewhat acceptable for volumes it is definitely not correct
with /tmp

Change-Id: I7dae58b1e48f0f75908563b31e1821195b726133