scripts: remove util-linux dependencies: mount -o remount
mount -o remount still needs to give a device to the kernel, so mount
will look it up in /proc/mounts but a bug in busybox mount makes this
apparently fail from time to time.

The kernel doesn't actually use the device, so always pass in '-' to
avoid the problem. This also works with util-linux's mount, which also
no longer looks up the mount point in the mount table.

Change-Id: I5c7a77ad9fdcc1fe55a8ebee4c38b63a7b72eed1