scripts/pre_rootfs: fix preserve files for coreutils 9.2+
coreutils 9.2 changed how cp -n worked and makes the command fail if
any file is skipped, so this stopped working in alpine 3.18
(tests were unfortunately stuck on alpine 3.17 just because jenkins job
was not updated, and caught it now...)

In this particular case, cp -u (--update) is equivalent because update
skips file when timestamp is identical and we only care about skipping
redundant copies in recursive context.
This also works with busybox cp.

Change-Id: Ia05490a534e11bb76061643ddea2218fdcf97396
Link: https://lists.gnu.org/archive/html/bug-coreutils/2023-01/msg00069.html