scripts: allow skipping from env
since swupdate now embeds the scripts it's no longer easy to not run
the scripts, e.g. in order to install 'plain' non-mkswu SWU images

In particular this makes the `make tests` swupdate target at package
build time fail (because running as non-root user in a non-ABOS
environment), so give ourselves a way out that's easier than fiddling
with the sw-description.

Ideally we'd only try to run the scripts if we know they are included in
the SWU, but there is no easy way of checking that (in particular pre
script format changed so even checking the first file after
sw-description in the archive would be fragile), so this will allow
passing build tests...

Change-Id: I49b8bfd0c27f4eeb860174927ffa785ee95dbd7b