scripts/init: refuse to install new container in atmark images
Atmark now ships SWUs which include containers signed with the official
key.

We don't want users to go around installing these containers on systems
that already run an application as that would break them or possibly
grant them undue access:
 - for updates signed by atmark
 - that contains a container not installed yet
 - fail if there already are containers

Since that hardcodes atmark public keys at yet another place in yet
another format in the code, add a check to make sure that code is kept
up to date in the future.
The value to use if update is required will be printed in its error.
(Move another code check that was in scripts.sh together to avoid
checking multiple times for each shell)

Note that for manual installs, it's possible to force CONTAINER_CLEAR
manually by just setting the env var (as per mkswu_var), so we also
allow this case explicitly.

Change-Id: If6d34a386d6d4f520de0b46374702cba4320e196