scripts_post: also kill swupdate on POST_ACTION=wait
Armadillo twin's implementation will be easier if swupdate stops on WAIT action.
We don't want to have hawkbit's swupdate signal success to the server, so
kill swupdate like we do with other rebooting modes.

For wait we do not want to leave a shell script with sleep infintiy behind,
so stop as soon as the parent process died.
Share this with other reboot modes in a function and comment why we kill
swupdate in the first place.

Also, while we could kill any other swupdate process with pkill, we
actually know which swupdate is our parent through PPID so use that
instead (we need it to wait anyway, it doesn't make sense to use pkill
and check PPID afterwards)

Change-Id: I80f606d6eaa334ae7df0fcf83d31eefac70a3fe9