scripts/versions: improve 'boot' 0-padding version handling
Since swupdate's if-different version handling treats 2020.04 and 2020.4
as identical versions we've been simplifying the version in mkswu and
install disk, but it seems the version can still be stored as 2020.04 in
/etc/sw-versions for some reason.

Until these bugs all are fixed we should make sure the version isn't
correctly detected as updated by swupdate as that could leave updates
unable to install or out of date on these systems.
The new workaround specifically fixes 'boot 2020.04' into 'boot 2020.4'
on both the system's sw-versions (if system is out of date) or in
sw-description (for old SWUs with new scripts); and the resulting
sw-versions after install will always have the 0 taken out.

Also, check for nothing to do in versions files using the fixed up boot
version so we correctly detect that there is nothing to update in case
the bad boot image was the only update candidate.
We can also simplify the nothing to do check and use cmp instead of
diff, as we haven't been filtering out boot_other for a while.

Link: https://armadillo.atmark-techno.com/forum/armadillo/17500
Change-Id: If4ee71966ebeae483f49504e960d118126f54280