hawkbit-compose: fix for new hawkbit-compose container
previous fix in 5.4 was incomplete, there were other changes:
- The hawkbit service no longer runs in /opt/hawkbit, so
application.properties file placed there was ignored.
Similarily, the artifactrepo also moved to /: fix the volume
mounts to use the proper path.
- The hawkbit service used to run as root, but now runs as a regular
user within the container, and can no longer write in the artifactrepo
if it comes from a volume mount.
Unfortunately, the hawkbit container does not have an entrypoint that
would fix permissions for it unlike mysql's: do a one-time chown after
container update if required.
- Hawkbit does not seem to run with our application.properties,
in particular 'spring.main.allow-bean-definition-overriding=true'
apparently needs to be set... Copy a bunch of default configuration
options from the default config and link to an up to date version
for further updates.

With this, hawkbit starts properly and both device registration and
updates (with and without rollouts) have been tested.

Change-Id: I868eda98269b85735d4e293cdcdeecf6ca018caf