scripts/pre_appfs: print mounted filesystems if btrfs subvol sync times out
btrfs subvol sync can timeout if deleted subvolumes are still mounted.
This shouldn't ever happen but we get such a report from time to time,
having the output of findmnt in this case would be helpful: just run it
and dump the few lines to info.

Since the volumes need to be umounted in all namespaces iterate through
all the mount namespaces we can find.

Change-Id: Id9cd9717892d1aa63d7e690cad93ac83f49e013e