scripts/pre_appfs: CONTAINER_CLEAR: also remount in pid1 namespace
When swupdate is started through udev's swupdate-usb, it runs in a
different mount namespace in order to isolate the disk's partition to
avoid conflicts with the rest of the base OS.

Unfortunately, it also means that the umount/mount logic we have here is
not enough: pid1's mount namespace still has all the old volumes
mounted.

Check if we run in a different namespace and if so use nsenter to also
remount volumes there.

Change-Id: Iffb48488af5c5ce289ee335747753ab284f73032