scripts/post fail action: setup fail action before running start command
We might not run either success nor fail command if setting up for the
fail command execution errored out.

If the user relies on either success or fail to always happen after
start that might be a problem, so prepare for fail command first.

Note if start command fails the update is also aborted, if running the
fail command when start command failed is a problem the user could add a
marker in start command and check for it in fail command, but that is
not expected to be a problem in practice. The opposite is not possible.

Change-Id: Ia4e696f5ba221faee112748bb031d1fb10e89a1d