scripts/common/lock: ensure lock is taken by root
This lock is now redundant with the lock taken by swupdate 2023.12, but we
must keep it for a while for backwards compatibility.

Thanksfully it's easy enough to check the lock is root-owned as
single-component mkdir is not susceptible to symlink attacks and
protects us from overwriting something with the PPID file.

Change-Id: Ia141207ac68cc1aa593fc713770c606c6eb4ae88