scripts/podman_*: move temporary directory to $TMPDIR/scripts
podman_cleanup and podman_target create a temporary directory which
should be cleaned up in normal operation, but if swupdate or the script
is hard-killed for some reason they might leave files behind.

$TMPDIR/scripts is $MKSWU_TMP which is guaranteed to be cleaned up the
next time an update starts (as well as when update is over in the normal
case), so at the very least this will allow these directories to not
pile up if interrupted multiple times.

Change-Id: Ia0945ed1cd7147aea3e383985cc3bbef5385ef48