scripts/pre_boot: add workaround for old Armadillo X2 eMMC firmware
older firmware for Armadillo X2/IoT G4 can brick the MMC if power is
lost while writing to boot partitions with >= 16KB blocks:
limit to 8KB if required.

Note this will also be set on boot, but older systems won't have the
setting and could still use the workaround, so set it on update too.

Change-Id: Ib982526042fab7d6562c4d36795118c4629ac25b