locales: remove built locales from repo
Having binary files in the repo makes it difficult for multiple patches
to touch locales in parallel, and is prone to forgetting to update it.

Let `make` be run at release time only, the check will still catch any
update to the po files (e.g. any locale forgotten to be updated) but git
will be able to merge these properly most of the time.

In order to bring the least friction possible, the built locales are
included in dist tarball: only people affected would be people running
mkswu from git, who need to run `make` to get localized messages if
required.

Change-Id: I0f2174c45a5e79771e04d1d2812a9a760f31155a