scripts/post: do not use 'touch -h'
We used 'touch -h' because for /tmp, if a user creates a symlink
/tmp/.swupdate_rebooting and we touch it we would end up creating the
symlink target as root with a regular 'touch'.

Unfortunately, touch -h has two problems:
- coreutils' touch -h will just fail if the target does not exist (!)
There is a --no-create option separate from --no-deference, but it
would seem that not following links also imply no create for some
reason...
- busybox touch does work correctly, but it does not use O_NOFOLLOW when
actually creating the file (utimensat fails -> open without NOFOLLOW),
so an attacker could create the link between the two syscalls.

This could easily be fixed in each implementation independantly, but
mkswu needs to work on old software, so just use something else:
- use regular touch for /run where this problem does not exist
- use mktemp + mv for /tmp, so mktemp does its job and mv's rename
overwrites whatever was there independantly.

We will remove this code next year anyway, so it does not matter if it
is a bit more complex than it needs to be.

Change-Id: Id140ec564ea9d460a8c4571e514e451f5a4fcd47