mkswu cleanup dir: fix cleanup with single quotes or spaces
This commit converts the code to use \0 for cleanup and fixes the
following problems:
- using newlines as separator means if files include new lines they'd be
incorrect, this was ignored on purpose as unlikely and not causing any
error (rm -f would just ignore them)
- using spaces would fail the join; this would make no path match as
it'd only match up to the space and filter out everything; this was not
intended but also is fine.
- using single quotes would be processed by xargs, and fail horribly.

That would also be fine if we did not have -e set here, it will be
removed in a later commit.

Link: https://armadillo.atmark-techno.com/forum/armadillo/21277
Change-Id: Icdefabb75665863a415685ccf906d9b441d9c50c