mkswu: add password strength check for Linux users
As part of the security authentication support for IoT devices, a
password strength check for Linux user passwords set during `mkswu
--init` has been added.

The cracklib-check command is used to check the password strength.

However, the cracklib-check command judges that passwords are okay as
long as they are at least 6 characters in length, so a separate check
is provided to ensure that they are at least 8 characters long.

The limit of 8 characters or more is in accordance with the Ministry
of Economy, Trade and Industry's ‘Security Conformity Assessment
System for IoT Products’ (☆1).

Redmine: 64054
Change-Id: Ic96025bf91fe0438ce416b125eef0fc010a833fc