scripts/cleanup: do not remove $MKSWU_TMP
swupdate started printing an error if the directory is missing when
trying to do its own cleanup, just leave the cleanup to swupdate.

This was already cleaned up properly in older versions of swupdate so we
do not need a conditional on the version; for older versions the lock
will be released before cleanup but using the same TMPDIR would crash
both updates anyway as waiting update would have already extracted
things before lock -- we don't need to try to make this work.

Note we keep the rm -rf at scripts_pre.sh: if a previous swupdate run
was interrupted with ctrl-C then we might have some garbage left behind
and we need to clean it up.

Change-Id: Iab0e6358081197170c6f782596450aa90a52e1d5