scripts: fail action: fix running fail action on pre-script failure
The fix to make fail action properly run from config on failure did not
properly run it if failure happened during setup (which should never
happen as that means updates are not installable, but can still happen
in theory)

We do not want to run the fail script before updates really started,
e.g. if updates have been rejected because of version check, so:
- remember that update really started in a new MKSWU_TMP file
- stop checking for the old lock, error() would free the lock and make
  cleanup.sh skip this incorrectly
- check the new file instead

also make it so that upgrade_available is cleared first to avoid
rollback; it's unlikely but possible that the start action makes the
target system unreliable so we should not rollback to it before this.

Fixes: 6fe114c78445 ("scripts/post fail action: run immediately on failure")
Change-Id: I48e6d9cd7d25e35a302529cd34c045d2455303c4