scripts/check_shadow_empty: allow empty passwords if expired
For users selling device with no password set (prompt to set on first
login) then we should allow users with no password.

If the password already is expired, allow the user to login.

Note that this would allow installs with the default root empty
password, which was what we wanted to prevent originally, but users can
only install SWUs after initial setup and initial setup always sets that
password, so that check is no longer required in this context.

Change-Id: Iba3d68d295aa3e4a86aac4fd677c5e28afa776df