mkswu: standardize MKSWU_desc_extra_text variable
the next commit will want to set MKSWU_desc_extra_text conditionally
in a subfunction.
It would be possible to make it local for just the caller, but if
a new caller shows up they risk leaking the variable which will
then be added to all following swdesc_* items.
Instead, add the variable to the local stanzas so we can use it
safely accross swdesc_* boundaries:
- calls from desc files will never leak the variable
- nested calls will work with the export logic

Change-Id: I0939dff7f8d91d330e9db61e3125c248079d991e