swdesc_boot: change default to install-if=higher
boot used to be install-if=different to allow for easier rollback in case
of problem, but in practice that is a security problem (easily revert to
vulnerable versions) and have not been useful in real situations.

If a revert is required, removing boot from sw-versions is easy enough.

Unfortunately, the default boot version can be quite convoluted: ensure
the default versions pass mostly ok.

Change-Id: I557e913126b809d1672401a3378379fb2e0bd628